Adding QueueAffinityStrategy.
This let's us ensure that items that are queued can define a key that
causes all items with that key to get queued to the same place. That way
we can ensure some serialization.

Right now this is the first pass and gradually over time the maps used to
decide which queue will grow and grow which is bad for large key sets.