Nuking the separate IMassTransitUriFactory's and making a single one.
It uses the TransportType to decide which sub factory to use. This way we
can just use this all the time, if somebody makes their own transport
then they can provide a whole new factory. Most of the time we'll be
using one of the supported ones and so this works better for most cases.