Added ParentContainer plugin that we can use to chain two containers.
The premise is that any dependency that can't be found in the child
is then looked for in the parent (by ServiceEntry.ServiceType.) and
used to resolve. The parent container is used exactly as if the
dependency were resolved manually and injected via overrides for example.
No special digging into internals is done.