Fixing a bug with ObjectFactories.
It didn't get to evolve with the container and so fell behind. One thing
we added was instance tracking and in order for that to work we can't
go around creating separate activators for the same entries. Another
complication was that we created this ServiceEntry on the fly every
resolution and so it would cause the final ResolvedServiceEntry's to be
inequal. To fix this, we just register the activator and entry's with
the system. This means we will only do this resolve step once for each
factory that's registered, which is nice too.