Started on IServiceCollection, a simple group of to-be registered services.
I find myself just registering types and in order to simplify and make
that process stand out and to group them easier I'm going to make it
more explicit in the container's domain. IServiceCollection is just meant
to register a collection of services and as such only takes a
ContainerRegister.