AddServiceCollection code in ContainerRegistrationHelper.
Code to iterate over an assembly and add all of the service collections
inside. That means adding them to the container and resolving them so
that they can have dependencies.

The only real issue here is if we need to specify some kind of order. Not
sure how I would like to do that just yet. I suppose people can just
inject them and decide themselves.