Getting Machine.MsMvc working with the new libraries.
I deleted a bunch of code having to deal with demultiplexing views from
an available collection of view engines. A lot has changed, and there even
seems to be an IViewEngine that does something like this. I chose to just
nuke this for now, we can choose a single view engine. Also, with Wepid
I'm wary of spending too much time on this.