Can Resolve with a dictionary of parameters arguments now.
This was a largish addition because I wanted to avoid a hack of some kind
and the best way to do this seemed to be to adjust the model starting with
how things are resolved. I needed the flexibility to get the key/string and
wanted to be careful about how I went around creating them in the model so
I abstracted their creation behind a ResolvableType object that produces
the ServiceEntry that has the key populated.

The resolving code is aware of the context that the resolve is happening
in and can pretty easily create a ResolvableType via the IResolutionServices
interface that it's got.

We use a custom IOverrideLookup to provide the values from the dictionary.

We could easily make a custom ServiceEntry that had a Key and see if
the thing we have inside of that lookup has a Key and use that in order to
keep people from using the Key in a weird way (the way it's used in many
other IoC's that I just don't find a need for)