Adding my interfaces as messages code.
This is the code we'll be using to get our initial support for passing
interfaces as messages. The important thing in here is the seriailzer
since that's the tricky part.

We'll probably be changing how we create message to take more care of
their immutability, but for now this should get us up and running.