Started a WebRequest lifestyle.
Right now we just use the HttpContext to find an existing instance and
delegate to the base class (TransientLifestyle) to create one otherwise.

I'm debating if I wanna write the code necessary to do Deactivate. Windsor
uses a module that performs a release in the EndRequest. This kind of
thing is important for pooled resources and other cases where we wanna
ensure we aren't collecting reference.