Actually invoking the reply callback's when we dispatch.
The only huge problem we have with this is that if a reply never comes back
then we've leaked a value because the entry in the dictionary will never
get removed. Not sure the best way to handle this, some kind of timeout
or something? I don't think NSB even takes are of that scenario.