Fake publish.
For now this is actually going to serve my needs! Over the next few weeks
I'll figure out a good way to get this working. I'm thinking there's
gotta be an easy way to re-use MT's publish/subscribe infrastructure
which is why I don't wanna rush into implementing this. Especially since
for now we don't have that many events or autonomous components in
the system to listen to them.