Caching the ServiceEntry for a given name after the first lookup.
This should help lookups that happen repeatedly, which should be the
more common scenario. In order to ensure the first lookup happens as
fast as possible we'd need to have some kind of event or something that
invalidated the cache when an entry's name was changed or require the
named entries to be added and setup before we started the container. Some
solution along those lines.

I'm thinking this should be enough.