An approach to container.Reset support. Deactivates *ALL* instances.
This is more experimental for some ideas we have regarding testing. We
desired the ability to start a container's history over, w/o having to
recreate the container from the beginning and such. This effectively will
Deactivate all instances that have been created so far. It requires a
tracking policy capable of remembering them to be installed (usually is)